Solved issue #128
Moved readData function to fileSystem
